didalam class employee terdapat class address dapat kita opindah kan dalam 1 file class tersendiri
