+++ src/lib/Address.java
@@ -1,0 +1,15 @@
+package lib;
+
+public class Address {
+    private String street;
+    private String city;
+    private String postalCode;
+    private String country;
+
+    public Address(String street, String city, String postalCode, String country) {
+        this.street = street;
+        this.city = city;
+        this.postalCode = postalCode;
+        this.country = country;
+    }
+}
